backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlmodel==0.0.22
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.2.5
yfinance==0.2.55
requests==2.32.3
python-dotenv==1.1.0
httpx==0.28.1
alembic==1.14.1
mypy==1.15.0
ruff==0.9.9
black==24.2.0
